Use server-side apply to deploy
Server-side apply will resolve how to update resources on the server so
that the spec we desire is always enforced. The force flags allows us to
override values if conflicts occur, claiming back ownership of a field.
API Services do not currently use SSA due to a bug that should already
be resolved in more recent Kubernetes revision versions.

Signed-off-by: Jakob Gray <20209054+JakobGray@users.noreply.github.com>