Hive SHA Hotfix for 2.10.0 (#2153)
The hive SHA cb63336e1595f38d68af721f9a64adb56f0b1d4e is too old to work
properly with the HIVE image that is packaged in the release. In order
to keep in sync with the Hive image, this PR updates the chart automation
to pull the CRDs and other resources from the same Hive commit.

Signed-off-by: Nathaniel Graham <ngraham@redhat.com>